+++ src/Pwa/Command/GenerateManifestJsonsCommand.mjs
@@ -1,29 +1,29 @@
 import { readFile, writeFile } from "node:fs/promises";
 
-/** @typedef {import("../../../../../flux-localization-api/src/Adapter/Api/LocalizationApi.mjs").LocalizationApi} LocalizationApi */
+/** @typedef {import("../../../../flux-localization-api/src/FluxLocalizationApi.mjs").FluxLocalizationApi} FluxLocalizationApi */
 
 export class GenerateManifestJsonsCommand {
     /**
-     * @type {LocalizationApi}
+     * @type {FluxLocalizationApi}
      */
-    #localization_api;
+    #flux_localization_api;
 
     /**
-     * @param {LocalizationApi} localization_api
+     * @param {FluxLocalizationApi} flux_localization_api
      * @returns {GenerateManifestJsonsCommand}
      */
-    static new(localization_api) {
+    static new(flux_localization_api) {
         return new this(
-            localization_api
+            flux_localization_api
         );
     }
 
     /**
-     * @param {LocalizationApi} localization_api
+     * @param {FluxLocalizationApi} flux_localization_api
      * @private
      */
-    constructor(localization_api) {
-        this.#localization_api = localization_api;
+    constructor(flux_localization_api) {
+        this.#flux_localization_api = flux_localization_api;
     }
 
     /**
@@ -34,37 +34,37 @@ export class GenerateManifestJsonsCommand {
     async generateManifestJsons(manifest_json_file, localization_folder) {
         const manifest = JSON.parse(await readFile(manifest_json_file, "utf8"));
 
-        await this.#localization_api.addModule(
+        await this.#flux_localization_api.addModule(
             localization_folder
         );
 
         const manifest_json_file_dot_pos = manifest_json_file.lastIndexOf(".");
 
-        for (const language of Object.keys((await this.#localization_api.getLanguages()).all)) {
+        for (const language of Object.keys((await this.#flux_localization_api.getLanguages()).all)) {
             const localized_manifest = structuredClone(manifest);
 
-            localized_manifest.description = await this.#localization_api.translate(
+            localized_manifest.description = await this.#flux_localization_api.translate(
                 localized_manifest.description ?? "",
                 null,
                 null,
                 language
             );
 
-            localized_manifest.dir = (await this.#localization_api.getLanguage(
+            localized_manifest.dir = (await this.#flux_localization_api.getLanguage(
                 null,
                 language
             )).direction;
 
             localized_manifest.lang = language;
 
-            localized_manifest.name = await this.#localization_api.translate(
+            localized_manifest.name = await this.#flux_localization_api.translate(
                 localized_manifest.name ?? "",
                 null,
                 null,
                 language
             );
 
-            localized_manifest.short_name = await this.#localization_api.translate(
+            localized_manifest.short_name = await this.#flux_localization_api.translate(
                 localized_manifest.short_name ?? "",
                 null,
                 null,